refactor: `src/modules` 내 entites 파일 도메인 폴더로 이동
