+++ src/common/entities/index.ts
@@ -1,0 +1,1 @@
+export * from './BaseEntity.entity';
