+++ src/modules/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
-import { User } from '@/models/entities';
+import { User } from '@/modules/user/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
