+++ src/modules/problem/entities/index.ts
@@ -1,4 +1,2 @@
-export * from './User.entity';
-export * from './SubmitCode.entity';
 export * from './Problem.entity';
 export * from './TestCase.entity';
