+++ src/modules/problem/problem.module.ts
@@ -2,7 +2,10 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ProblemController } from './problem.controller';
 import { ProblemService } from './problem.service';
-import { Problem, SubmitCode, User, TestCase } from '@/models/entities';
+import { Problem } from '@/modules/problem/entities';
+import { TestCase } from '@/modules/problem/entities';
+import { SubmitCode } from '@/modules/submit/entities';
+import { User } from '@/modules/user/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Problem, SubmitCode, User, TestCase])],
