+++ src/modules/submit/submit.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { SubmitController } from './submit.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { SubmitCode } from '@/models/entities';
+import { SubmitCode } from '@/modules/submit/entities';
 import { SubmitService } from './submit.service';
 
 @Module({
