+++ src/modules/user/user.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
-import { User } from '@/models/entities';
+import { User } from '@/modules/user/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
