+++ src/modules/user/user.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { User } from '@/models/entities';
+import { User } from '@/modules/user/entities';
 import { Repository } from 'typeorm';
 
 type UserInfo = {
